requirements_dev.txt
-r requirements.txt

apache-airflow-providers-postgres==5.6.1
black